Fix bug introduced in WordPress 4.2.3.
It appears this commit may have been committed in 2013, but wasn't released until now. In any case, WordPress 4.2.3 added some sanity checks in it's meta cap checks that meant, instead of returning `true` on `current_user_can` when passing in 0 as the post_id, it now returns false. This updates the behavior to check if the user can edit_posts generally if we have a post_id of 0 (which means the post hasn't been created yet).