Implement `/gistpen/search` endpoint
Start breaking down Api\Ajax and removing unncessary logic. A
lot of this is handled by WP-API, and much of the rest will be
handled by the framework.

This mostly wires up the current services. Once the old AJAX
endpoints are wired up, we'll start migrating the old services into
the framework.