Wire up zip controller's view and create methods
Adds a GET and POST route for retrieving and creating Zips,
respectively. Also introduces a ZipFilter for validating and
sanitizing the Zip and File data.

Some of this logic will pulled from the old Save controller as well
as the Ajax API endpoint. This commit removes that logic from
their current locations.