Unit test provider instantiation
No need to fully unit test the providers, I don't think, but this
is a good way to ensure the providers are all being loaded as
expected. This also provides some coverage on the filters/actions,
as the framework throws if those are misconfigured, so those
errors will be caught too.