Update UserControllerTest to extend from TestCase
Also update the UserController to return the invalid keys in the
headers, rather than adding an extra key in the response. This
allows us to maintain the response as containing only the data
retrieved from the API endpoint, while still providing the necessary
information to inform the client to not use the invalid keys.