Harmonize code snippet display
Remove the header with the code snippet's file name and
move that information to a bubble in the highlighted snippet.
This also removes the show language bubble, since the
file name has enough information to know the language.
Simplifying the UX is definitely a plus, so removing this
option and making it look the same everywhere ensures some
consistency.