Eliminate dead code for React settings page
The Settings page just echos a div React is going to mout to. Now,
we can elimiate CMB2 completely, plus extra tests we don't need.
The Gist client's tests are going to need to rewritten too, as we
want to get off of `php-github-api`, so we need to figure out another
way to supply the token, rather than fetching it directly out of CMB2.