Rewrite build script to user webpack
We'll migrate the styles loading to webpack as well, to
simplify the build scripts overall. Gulp right now primarily
handles the copying of the library files and managing
the config for webpack, which handles building the
JS and (soon) the CSS. We can then flatten the folder
structure a bit and make the FE development experience
cleaner.

Includes forcing translation regneration in development flow.

Fix #51.