Performance enhancements in settings page
If a server update is in progress, we cache the patch and
redispatchthe AJAX call with the combined patch. We can
also abort reinjecting the iframe if the value of the site
settings haven't changed since the last time we injected it.
Finally, we can do a deep equality check to see if the state
has changed and only reupdate the state if it's changed.