Rewrite application with streams for flow control
Create streams for all the events coming out of the UI and
merge them all together for piping through a patcher (which
functions similar to a reducer) to create patch objects.
These patches are passed into the store to mutate the state,
which it then emits to the subscribed render function.

We may not end up using rx-react long-term, since we've
just got the single function we're using from it, which can
be handled internally.