Document new endpoints in API Blueprint
This lays out the resource structure for the new WP-Gistpen
API, giving us enough information to design the system from
end-to-end, including model data structures and relationships.
We can use this to flesh out both framework- and application-
level constructs to get the thing up and running, and bring
the framework up to the point where it's actually useful for
building plugins.

If followed, it will resolve the concerns addressed in #56.