Move enhancer creation into app bootstrap
`delta` folder is just a collection of delta or delta-creating
functions. Binding them together into the Redux enhancer is
handled in the individual applications' bootstrapping function.
Also delete the middleware folder, since no middleware exist
yet and they will be handled the same way as deltas.

Also refactors the routerDelta into a createRouterDelta function
so it can be bound to any router that uses a hash.