Rewrite web JavaScript & Prism interface for async
A lot of the functionality we implemented previously to async
load the CSS & JavaScript can be deferred to being handled
by webpack. While I'm not in love with the idea of loading the
JS in the head and using that to inject the styles, it does make
for a really clean implementation, allowing all the dependencies
to be managed in the JavaScript, rather than coordinating
multiple systems.

Shims are added for all the plugins & themes, which allows
us to load them all with a System.import expression. The Prism
interface also loads up the toolbar plugin for us as well as the
buttons registered with it. This allows us also to clean up the
`copy` step and extract it.