Initial editor implementation copied from Dabblet
Big thanks to Lea Verou for her work on the editor in Dabblet. This uses
a lot of that code as the starting point for building out our own Prism-
based editor. There's still a lot of cleanup to be done to push it into the
architecture, but this gets a big chunk of the functionality implemented.