Wrap code in `code` element
The `code` element, instead of the `pre` element, will be
`contenteditable`. This allows us to use the `line-numbers` and
`line-highlight` plugins from Prism to handle those display elements,
and push off some complexity out of our editor. The selection-related
functions will have to be extracted, but that will be handled when we move the `keyup` logic into the `render` function.