Extract render logic to render runfction
This eliminates a lot of unncessary bespoke logic from the `onMount`
function and pushes it into the `render` function, which ensures the
Prism settings are updated correctly before updating the element to
match the template, then highlighting the element with Prism. Finally,
we'll move the cursor setting logic here as well, eventually.