Add debounce to ensure latest value
This allows us to capture the element after the events effects has
occured so we can pull the value from the element with its new value,
instead of needing to reconstruct what happened in the reducer. We
may want to actually remove this and do exactly that, though, if we
run into problems.