Move rendering to onMount and lean on input event
Instead of trying to trap particular keypresses and interpret them into
the new editor value, we're going to lean much more heavily on the
`input` event to give us the latest value, saving the previous value and
the previous cursor state in the history. This does mean we need to
track the cursor as it moves around, so we can reset it, so the next
steps in the implementation will be to ensure we capture all the cursor
moves and save that information in the state.

From there, a handful of commands will need to be interpreted,
including indentation, transforming a comment, and adding a new line.
These will be interpreted in the reducer, creating a new state for the
editor to render. We'll have to see how well this works with the current
rendering strategy, but this should get us most of the way there.