Add basic AVA setup w/ gulp-ava
The gulp task is responsible for setting the BABEL_ENV so babel
compiles the modules correctly. A `dev:tdd` task has been added
to retest regularly, which seems to be working smoothly.

Comes with a basic test so AVA can find something to test, but now we
can start writing units tests as appropriate.