Add initial logic for make newline
Uses the indentation from the previous link to indicate where the tab
should be added on the new line. This could be made smarter to do
some neat things like auto-add curly brackets, add extra indentation,
etc. However, that may need to be handled per language, which could
be prohibative. Regardless, this provides an initial implementation to
handle adding newlines.