Fix edit toolbar styles and theme dropdown
Using a standard HTML dropdown with some slight custom styling.
`white-space` was the property causing the issue. Once I found that
`all: initial` would kinda work, I stepped through all of the properties
and reset of them to initial with this script:

```js
(() => {
    const toolbar = document.querySelector('#post > div.wpgp-editor > div:nth-child(3) > div > pre > div');

    Object.keys(toolbar.style).filter(prop => prop !== 'all').forEach(property => {
		toolbar.style[property] = 'initial';
		debugger;
    });
})()
```

Execution paused on each reset, allowing me to check which property,
when reset to `'initial'`, fixed the toolbar.