Add support for `userDelta` to update editor opts
Add additional ajax-related actions to distinguish success per endpoint
rather than generically against the API. Update the reducers to use
this new action type, and use the new action to dispatch after the
user settings are saved (the editor controls). Lots of new types for
all of these things, and everything continues to pass flow/tcomb.

We have to figure out how to use the strings in types for tcomb, but
we have an open question here:

https://github.com/gcanti/babel-plugin-tcomb/issues/161

No tests in place, unfortunately :(.