Mount root onto the body
This allows the root to automatically mount any relevant components
whenever they enter the DOM. This means we can just inject the component
wherever we need it, and it'll automatically get wired up, start emitting
events, and keep its portion of the DOM in sync with the props, as
appropriate. Because the root component doesn't have a `render` function
it doesn't attempt to rerender the whole body.

Kinda nifty, huh?