Update settings to use query string routing
The new router works by reading the query string and checking if
the provied param exists. If not, we route to the default route with
`pushState` and dispatch a route change. It listens to href clicks
and history changes and we also dispatch route changes that way,
keeping the route & state in sync with `pushState`.