Add Job Contract
This assumes the Job is responsible for managing its entire
lifecylce, kicking itself off, handling the processing, and
continuing itself if it doesn't finish in a single process.
This is probably not ideal long-term, as we should move
some of this handling out of an individual Job and into the
supporting cast, but we can put that functionality in an
AbstractJob we extend from for now and break those
pieces out into separate classes later. It doesn't mean the
Job cannot be reused outside of Gistpen until that's settled
but right now we just want to get the functionality working.