Add Jobs\Manager & Service Provider
This is pretty basic, as all it currently does it wrap a
Dictionary, but since we are eventually going to have to
move things up a level into the "supporting cast", this
will be a site we can move things, so this is fine for now.
It also registers the Manager so we can start pulling it into
things.