Implement AbstractJob
This simplifies the process of creating a new job to
implementing 4 methods:

* name
* description
* fetch_items
* process_item

This keeps the scope of what needs to be handled in
userland to a focused minimum on how to fetch &
process the items, while leaving everything to the
AbstractJob to manage the overall flow. Everything is
implemented as tightly as possible; most of the variables
and methods are private, so the implementing class only
has access to `log`, `get_timestamp`, and the current
EntityManager implementation. This allows us to break
out the responsibilities of the AbstractJob into other
classes as we need them to enable unit testing, moving
this functionality into isolated classes we can mock.