Add Jobs params and provide minimal data
Depending on what route the user is on when the page
loads, the Jobs params provides the minimum amount
of data needed for the page to render. Additional data will
need to be fetched as the user moves around.

This requires the routers to be updated to pull that data.