Create Gist client service
This is responsible for sending HTTP requests. I wrote a
small script that I ran with wp eval-file test-client.php,
which gave me access to my site's internals so I could
just use the token and implementation there. I don't plan
on setting this up via CI, but it is useful to check that it
does what we think it does. Unit testing the client would
be the next real step.