Implement ExportJob to export Repo to Gist
Right now, we're not taking advantage of the reprocessing capabilties
if something goes wrong. Running these things over and over, and
logging all of the time, seems like a terrible idea, although it's
certainly possible to do. Is this something we should even make
available?

There are reasons to reprocess an item, so it may be worth giving
users the power to reprocess if they choose, but it is an awesome
power to be used wisely, as I have triggered crazy loops where 5k+
messages were sent to the log table while developing this.

But once it was implemented, the code actually comes out looking
pretty clean, all things considered.