Convert `flatten` to normal `map`
`flatten` always returns a stream, and the render logic in the
instance's `onMount` is assuming `props$` is a property, so we
can take the first value from it and render the editor based on it.
Because that assumption doesn't hold, the editor gets added to
the page but not rendered by the instance until the next state
change.