Add initial support for dredd
This gives us integrated API documentation as well as tests where we
can validate that the API responds with the types we have defined
in our application. We can use tcomb or flow-runtime to check the
types in the tests and confirm they match that as well.

If both this and the previous commit get fully fleshed out, we can
validate that the entire system functions correctly.