Fix deleting bug
This query doesn't work correctly yet (see #113), so we need
to pass in more options to make it work. We're just matching
what `fill_relations` does for now, and we can update this
later.

What was actually happening was, whenever WP goes to clean
out auto-drafts, we would use this hook to make sure we delete
the repo's blobs. The problem is the query is too expansive,
pulling in blobs that don't belong to the repo, so we'd end up
deleting a bunch of blobs we shouldn't. This changes ensures
we have the correct results for the delete query.