Add Translator service for front-end templates
Extract all of the translations to a Translator service
and render them in the header of the admin so the
JS can pull it in and use it to render on the front-end.
This gets the BE to render correctly with the translations
and pick up with it on the front-end.