Change up codestyle and make Flow do linting
All of the type-related lints should be handled by Flow, which
can give us better insight into how well we can rely on the
typechecker. The eslint config is just for code style. This
also updates everything for the new style. Biggest change
is removing the space before the type `:`.