Fixed some errors in API handling
Better off creating a specific type for the API request, as the Blob response
and everywhere else should have an ID property. Making the types scoped
to the areas we have to deal with them makes more sense, so the API
request type is a separate type from the Blob domain model.

Also fixes up the `mapErrors` -> `flatMapErrors`, so we get an actual
Action emitted into the store rather than an error.