Integrate PHP-DI
We're going to move things around a _little bit_, but it will be nice
to get some autowiring and defining things without writing up a
whole Provider for it. Break up the configuration into a few files,
and move a few things into `resources`. Update the tests to
make the instance in the container. Add webpack alias for
resources folder.