Refactor out old container
This is a pretty large refactor, but it completes the transition
into PHP-DI. We remove all of the providors in favor of
configration, so we can eventually set up caching for the
container which we generate on deploy. We also remove
the Klass enum & various constants, not that we can use
the `class` static constant.

We also simplify the `boot` & `container` functions, as
these can function as the bare glue code needed to bind
the app to WordPress. There's definitely some cleanup
that can be done around how assets and routes are
bound into the container (we're kind of cheating by
passing in the Register services to `register` because
they just need to be created and have those methods run,
not anything else).