Extract BlobCreate filter & remove custom impl (#860)
A lot of the rules we've implemented with custom methods can
instead be handled by WordPress core. We can instead remove
the methods and use the normal schema logic.

We also extract BlobCreate to a separate filter and borrow its
rules in the RepoCreate filter so we can reuse them across both
the Repo creation controller and incoming Blob creation controller.

Fix #614.