Use Prism directly with babel plugin
This cleans up the implementation a bit. Instead of extending Prism,
we provide extra functions for interacting with it that can be used
across the application. Now we bundle the autoloader via the babel
plugin instead of needing to manage it ourselves.

None of this should produce any functional change in the front-end
but should simplify our usage of Prism, keep it a little cleaner, while
making it easier to integrate into the editor.