Add EditEmbed component & wire up (#951)
This should complete the initial integration of the Gutenberg block.

A lot of deduplication here: we have actions that overlap the old
and new editors, so we either need to change the name of the
action or switch to using the new action. Since we're going to be
killing off the old editor eventually, we're going to give those bad
names for now.

Fixes #278.